ai/requirements.txt
fastapi
uvicorn
requests
asyncpg
httpx
python-dotenv
redis
pydantic
